update Qt version to Qt6
Changes to various files in src:
- explicitly cast the result of Qt .length() results
  In Qt6 the return type of those methods is qsizetype, which on most relevant
  platforms should be identical to ssize_t.
  This commit is part of the preparation for updating to Qt6.
- remove #ifs regarding Qt5 versions
  The newest Qt5 version it checked for was 5.6 which was released over 9 years
  ago and with the update to Qt6 this makes all of these checks obsolete.

cmake/CollectWindows: add newly missing dlls

COMPILE.md:
  - update dependencies
  - remove references to Ubuntu 20.04,
  - remove references to custom Qt installation paths
  - update Windows mingw set up

+++ src/amun/strategy/typescript/node/buffer/buffer.cpp
@@ -105,7 +105,7 @@ void Node::Buffer::Instance::lengthGet(Local<String> property, const PropertyCal
     Local<External> wrap = info.Holder()->GetInternalField(Node::Buffer::Instance::OBJECT_INSTANCE_INDEX).As<External>();
     auto instance = static_cast<Node::Buffer::Instance*>(wrap->Value());
 
-    info.GetReturnValue().Set(instance->m_data.length());
+    info.GetReturnValue().Set(static_cast<int32_t>(instance->m_data.length()));
 }
 
 void Node::Buffer::Instance::toString(const FunctionCallbackInfo<Value>& args) {
